Structured all the python modules, beautified and commented the files
